requirements.txt
aws_cdk.core
aws_cdk.aws-dynamodb
aws_cdk.aws-apigateway
aws_cdk.aws-lambda
